LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-text-to-the-canvas/2d.text.measure.actualBoundingBox-expected.txt

2d.text.measure.actualBoundingBox
Testing actualBoundingBox
Actual output:

FAIL Testing actualBoundingBox assert_equals: Math.abs(ctx.measureText('A').actualBoundingBoxDescent) === 0 (got 1[number], expected 0[number]) expected 0 but got 1
